+++ philo/utils_philo.c
@@ -12,7 +12,7 @@
 
 #include "philosophers.h"
 
-size_t	get_current_time(void)
+int	get_current_time(void)
 {
 	struct timeval	tv;
 
@@ -25,9 +25,9 @@ size_t	get_current_time(void)
 	return ((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
 
-void	precise_sleep(size_t milliseconds, t_table *table)
+void	precise_sleep(int milliseconds, t_table *table)
 {
-	size_t	start;
+	int	start;
 
 	(void)table;
 	start = get_current_time();
@@ -37,15 +37,15 @@ void	precise_sleep(size_t milliseconds, t_table *table)
 	}
 }
 
-void	print_status(t_table *table, size_t philosopher_id, const char *status)
+void	print_status(t_table *table, int philosopher_id, const char *status)
 {
-	size_t	current_time;
+	int	current_time;
 
 	pthread_mutex_lock(&table->print_mutex);
 	if (!table->simulation_stop)
 	{
 		current_time = get_current_time();
-		printf("%zu %zu %s\n", current_time - table->start_time, philosopher_id,
+		printf("%d %d %s\n", current_time - table->start_time, philosopher_id,
 			status);
 	}
 	pthread_mutex_unlock(&table->print_mutex);
